+++ CHANGELOG.md
@@ -11,8 +11,9 @@
 * Added `VERSION` value. Useful for browser dependency checking.
 * Added `lines()` method.
 * Added `slugify()` method. 
-* Added 'escapeHTML()' and 'unescapeHTML()' methods.
-* Added 'truncate()' method.
+* Added `escapeHTML()` and `unescapeHTML()` methods.
+* Added `truncate()` method.
+* Added `toCSV()` and `parseCSV()` methods.
 
 0.2.2 / 2012-09-20
 ------------------